refactor: 💡 extract injectToolbarStickyCSS to editorOperator
