+++ src/modules/createRankShortcut.js
@@ -1,4 +1,4 @@
-import { getEditor, getNameLink, insertToEditor } from '../utils/editorOperator'
+import { getEditor, getNameLink, insertToEditor, injectToolbarStickyCSS } from '../utils/editorOperator'
 
 const RANKS = {
   TRT_HARDER: {
@@ -82,21 +82,4 @@ function appendShortcutSelect (appendDom, id) {
     select.innerHTML += `<option value="${value}">${name}</option>`
   })
   appendDom.prepend(select)
-}
-
-function injectToolbarStickyCSS () {
-  const injectStyle = document.getElementById('helper-inject-css')
-  if (!injectStyle) {
-    const head = document.head || document.getElementsByTagName('head')[0]
-    const style = document.createElement('style')
-    style.id = 'helper-inject-css'
-    style.textContent = `
-      trix-toolbar {
-        position: sticky;
-        top: 65px;
-        z-index: 4;
-      }
-    `
-    head.appendChild(style)
-  }
 }
